// lib/generated/prisma/models/User.ts
export type UserModel = runtime.Types.Result.DefaultSelection<Prisma.$UserPayload>
// ...
export type $UserPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "User"
  objects: {
    team: Prisma.$TeamPayload<ExtArgs> | null
    managedTeams: Prisma.$TeamPayload<ExtArgs>[]
    createdProposals: Prisma.$ProposalPayload<ExtArgs>[]
    approvedProposals: Prisma.$ProposalPayload<ExtArgs>[]
    proposalVersions: Prisma.$ProposalVersionPayload<ExtArgs>[]
    approvalEvents: Prisma.$ApprovalEventPayload<ExtArgs>[]
    auditLogs: Prisma.$AuditLogPayload<ExtArgs>[]
    notifications: Prisma.$NotificationPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    email: string
    role: $Enums.Role
    jobTitle: string | null
    avatarUrl: string | null
    teamId: string | null
    defaultApproverId: string | null
    isActive: boolean
    createdAt: Date
    lastLoginAt: Date | null
  }, ExtArgs["result"]["user"]>
  composites: {}
}